# perf: transpose the PQ codes to improve search performance