… rust/lance/src/index/vector/builder.rs
… rust/lance/src/index/vector/ivf/io.rs
… rust/lance/src/index/vector/ivf/v2.rs
… rust/lance/src/index/vector/pq.rs